requirements.txt
playwright>=1.44,<2.0
pytest>=8.0,<9.0
pytest-playwright>=0.5.0,<1.0
python-dotenv>=1.0,<2.0
